add notification to block/unblock actions, edit PostmanCollection

+++ server/app/app/Http/Controllers/AdminController.php
@@ -3,34 +3,58 @@
 namespace App\Http\Controllers;
 
 use App\Constants\RoleConstants;
-use App\Http\Requests\UserIdRequest;
 use App\Http\Resources\BlockResource;
 use App\Http\Resources\UnblockResource;
 use App\Models\Role;
 use App\Models\User;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
+use App\Models\UserNotification;
+use Exception;
+
 
 class AdminController extends Controller
 {
     public function blockUser(User $user)
     {
-        $role = Role::query()->where('role', RoleConstants::USER_BLOCKED)->first();
-        $user->roles()->attach($role);
-        $user->save();
-        return new BlockResource($user);
+        if(!$user->hasRole(RoleConstants::USER_BLOCKED)) {
+            $role = Role::query()->where('role', RoleConstants::USER_BLOCKED)->first();
+            $user->roles()->attach($role);
+            $userId = $user->id;
+            $user->save();
+            $this->sendBlockNotifications($userId);
+
+            return new BlockResource($user);
+        }
+        throw new Exception("User already has been blocked", 402);
+
     }
 
     public function unblockUser(User $user)
     {
-        $role = Role::query()->where('role', RoleConstants::USER_BLOCKED)->first();
-        $user->roles()->detach($role);
-        $user->save();
-        return new UnblockResource($user);
+        if($user->hasRole(RoleConstants::USER_BLOCKED)) {
+            $role = Role::query()->where('role', RoleConstants::USER_BLOCKED)->first();
+            $user->roles()->detach($role);
+            $userId = $user->id;
+            $user->save();
+            $this->sendUnblockNotifications($userId);
+
+            return new UnblockResource($user);
+        }
+        throw new Exception("User already has been unblocked", 402);
     }
-    public function wishlist(User $user)
+    public function sendBlockNotifications($userId)
     {
-        $wishlists = $user->wishlists;
-        return response()->json(['places'=>$wishlists]);
+        UserNotification::query()->create([
+            'user_id' => $userId,
+            'type' => 'push',
+            'content' => 'User with id: ' . $userId . ', you have been blocked.'
+        ]);
+    }
+    public function sendUnblockNotifications($userId)
+    {
+            UserNotification::query()->create([
+                'user_id' => $userId,
+                'type' => 'push',
+                'content' => 'User with id: ' . $userId . ', you have been unblocked.'
+            ]);
     }
 }
